- Fixed Android warnings found with some upcoming toolchain changes - Allow GoogleARSDK for x64 architecture #rb chris.babcock #preflight 638a5343976b1644cb6d9c01
[CL 23377164 by Josh Adams in ue5-main branch]

+++ Engine/Source/Runtime/ApplicationCore/Private/Android/AndroidApplication.cpp
@@ -157,10 +157,10 @@ void FDisplayMetrics::RebuildDisplayMetrics( FDisplayMetrics& OutDisplayMetrics 
 	// Apply the debug safe zones
 	OutDisplayMetrics.ApplyDefaultSafeZones();
 
-	float Inset_Left = -1.0f;
-	float Inset_Top = -1.0f;
-	float Inset_Right = -1.0f;
-	float Inset_Bottom = -1.0f;
+	double Inset_Left = -1.0f;
+	double Inset_Top = -1.0f;
+	double Inset_Right = -1.0f;
+	double Inset_Bottom = -1.0f;
 
 	bool bIsPortrait = FAndroidWindow::IsPortraitOrientation();
